refactor: move SearchForm to SharedComponents
