+++ src/components/SharedComponents/SearchForm/SearchForm.stories.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import AnalyticsDecorator from "../../storybook-decorators/AnalyticsDecorator";
-
 import Component from "./SearchForm";
+
+import AnalyticsDecorator from "@/storybook-decorators/AnalyticsDecorator";
+
 
 export default {
   decorators: [AnalyticsDecorator],
